+++ Model/Reepaysepa.php
@@ -8,4 +8,9 @@ class Reepaysepa extends \Radarsofthouse\Reepay\Model\Reepaypayment
      * @var string
      */
     protected $_code = 'reepay_sepa';
+
+    /**
+     * @var boolean
+     */
+    protected $_isAutoCapture = true;
 }