+++ Model/Reepayswish.php
@@ -8,4 +8,9 @@ class Reepayswish extends \Radarsofthouse\Reepay\Model\Reepaypayment
      * @var string
      */
     protected $_code = 'reepay_swish';
+
+    /**
+     * @var boolean
+     */
+    protected $_isAutoCapture = true;
 }