+++ Model/Reepayverkkopankki.php
@@ -8,4 +8,9 @@ class Reepayverkkopankki extends \Radarsofthouse\Reepay\Model\Reepaypayment
      * @var string
      */
     protected $_code = 'reepay_verkkopankki';
+
+    /**
+     * @var boolean
+     */
+    protected $_isAutoCapture = true;
 }